nop bai tuan 6

+++ demo/src/main/java/com/example/demo/controller/BookController.java
@@ -8,8 +8,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
-import org.springframework.web.bind.annotation.*;
 import org.springframework.validation.Errors;
+import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
@@ -18,58 +18,66 @@ import java.util.List;
 public class BookController {
     @Autowired
     private BookService bookService;
+
     @Autowired
     private CategoryService categoryService;
+
     @GetMapping
     public String showAllBooks(Model model){
         List<Book> books = bookService.getAllBooks();
-        model.addAttribute("books",books);
+        model.addAttribute("books", books);
         return "book/list";
     }
+
     @GetMapping("/add")
     public String addBookForm(Model model){
         model.addAttribute("book", new Book());
-        model.addAttribute("categories",categoryService.getAllCategories());
+        model.addAttribute("categories", categoryService.getAllCategories());
         return "book/add";
     }
+
     @PostMapping("/add")
-    public String addBook(@ModelAttribute("book") @Valid Book book, BindingResult bindingResult, Model model){
-        if (bindingResult.hasErrors()) {
-            // Xử lý lỗi ở đây
+    public String addBook(@Valid @ModelAttribute("book") Book book, BindingResult bindingResult, Model model){
+        // Truong hop co loi rang buoc thi tra lai view add
+        if (bindingResult.hasErrors()){
             model.addAttribute("categories", categoryService.getAllCategories());
             return "book/add";
         }
+
         bookService.addBook(book);
         return "redirect:/books";
     }
+
     @GetMapping("/edit/{id}")
-    public String editBooks(@PathVariable("id") Long id, Model model){
-        for(Book book : bookService.getAllBooks()){
-            if (book.getId().equals(id)){
-                model.addAttribute("book", book);
-                model.addAttribute("categories", categoryService.getAllCategories());
-                return "book/edit";
-            }
+    public String editBookFrom(@PathVariable("id") Long id, Model model){
+        Book find = bookService.getBookById(id);
+        if(find != null){
+            model.addAttribute("book",find);
+            model.addAttribute("categories", categoryService.getAllCategories());
+            return "book/edit";
+        }else{
+            return "not-found";
         }
-        return "redirect:/books";
     }
 
     @PostMapping("/edit")
-    public String editBooks(@Valid Book updateBook,Errors errors, Model model) {
+    public String editBook(@Valid Book updateBook, Errors errors, Model model){
         if (errors != null && errors.getErrorCount() > 0) {
             model.addAttribute("categories", categoryService.getAllCategories());
             return "book/edit";
         }
         else
-        // Logic để xử lý dữ liệu sách đã chỉnh sửa và lưu vào cơ sở dữ liệu
-        bookService.updateBook(updateBook);
-
-        // Chuyển hướng người dùng đến trang danh sách sách
+            bookService.updateBook(updateBook);
         return "redirect:/books";
     }
+
     @GetMapping("/delete/{id}")
-    public String deleteBooks(@PathVariable("id") Long id) {
-        bookService.deleteBook(id);
-        return "redirect:/books";
+    public String deleteBooks(@PathVariable("id") Long id){
+        Book find = bookService.getBookById(id);
+        if(find!=null){
+            bookService.deleteBook(id);
+            return "redirect:/books";
+        }
+        return "not-found";
     }
 }
